[MIRROR] Adds a martial artist gi to the AutoDrobe. [MDB IGNORE] (#19692)
* Adds a martial artist gi to the AutoDrobe. (#73794)

## About The Pull Request
On a dark and stormy night, @  Cheshify asked in OOC if anyone would be
down to make a martial arts gi of no known origin for an event idea they
had.

Naturally, I decided to put my talent (CITATION NEEDED) to use, and I
have both sprited and coded in (the latter with the help of Cheshify
ofc) the aforementioned gi of no known origin.
<details>
<summary>Picture</summary>

![gokuss13](https://user-images.githubusercontent.com/117083785/222961961-a94c88fb-adae-4d2a-96c9-0f347e17c40b.png)

</details>

## Why It's Good For The Game

![image](https://user-images.githubusercontent.com/117083785/222961938-0f5c7508-0197-45f4-82e0-2d1ea2b2ee65.png)

## Changelog

:cl: dessysalta, Cheshify
add: Nanotrasen has added a martial artist gi to the AutoDrobe for those
who want to go even further byond.
imageadd: Aforementioned gi sprites.
/:cl:

* Adds a martial artist gi to the AutoDrobe.

---------

Co-authored-by: Rem/Dess <117083785+dessysalta@ users.noreply.github.com>