Fix inconsistency in FREEZE_PROOF flag [MDB IGNORE] (#19979)
* Fix inconsistency in FREEZE_PROOF flag (#74102)

Although all objects that _have_ the flag attach it to
`resistance_flags` (which is what the define location implies it's
supposed to be set on), all the code that checked it for the purposes of
applying the freezing element used `obj_flags`.

As far as I can tell, that meant the only things immune to freezing were
things that "blocked z falling from above".

Also, freezing only happens with some obscure weapons and low
temperature water vapour, but hey 🤷.

:cl: coiax
fix: Coffee cups are now correctly immune to becoming frozen by low
temperature water vapour.
/:cl:

Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>