Using bluespace launchpads with circuits now work again. [MDB IGNORE] (#20025)
* Using bluespace launchpads with circuits now work again. (#74134)

## About The Pull Request
Closes #74106
Makes it so that it actually uses the values of the object instead of
the object it self.
First PR so please bear with me. CharredPotato helped immensely with the
PR. I thank him for that.
### Mapping March
Ckey to receive rewards: N/A

## Why It's Good For The Game
bug bad
fix kill bug
fix good
## Changelog
:cl:
fix: Circuit-based bluespace launchpads now work again!
/:cl:

* Using bluespace launchpads with circuits now work again.

---------

Co-authored-by: Singul0 <127663818+Singul0@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>