Streamlines/Fixes Piracy Spawner Checks - They Piss Off When Paid [MDB IGNORE] (#20167)
* Streamlines/Fixes Piracy Spawner Checks - They Piss Off When Paid (#74275)

I think at some point this worked, but there were just too many checks
on abstract concepts and it seems that "aborting" the callback wasn't
really working? Anyways everything just felt really awk to me with how
`message_override` was working, so let's store a variable on
`chosen_gang` since we're already reliant on that for all of our
messages and stuff like that, and just slim down the number of weird
checks that we do into just one simple "did we get paid"?

One consequence of this is that refusing to pay won't spawn pirates
instantly, but the time-to-respond was already two minutes so I don't
really see this as a major drawback.
## Why It's Good For The Game

I PAID THEM OFF NOW FUCK OFF

* Streamlines/Fixes Piracy Spawner Checks - They Piss Off When Paid

---------

Co-authored-by: san7890 <the@san7890.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>