Removes fire cremation [MDB IGNORE] (#20152)
* Removes fire cremation (#74305)

Added in #44764

Removes the code that allows a human to be cremated by high burn damage
and being on fire

:cl:
del: removes being cremated from high burn damage and being on fire (yes
this was in the game but has been broken for probably a very long time)
/:cl:

### Why this is good for the game
This was added about 4 years ago, but never really worked (or broke
pretty quickly afterwards) because of a rounding error in the burn
damage code. GoldenAlpharex actually stumbled upon this when working on
a downstream and explained how it broke, and I kind of agree that it
doesnt fit that well into the game anymore

This was added when cloning was still a thing and being husked didn't
mean much if the cloner was upgraded. With the removal of cloner, being
husked is enough punishment in my opinion and we dont need to add
cremation. We've also gone without this feature for quite a while and I
haven't missed it

EDIT: Oops, just talked with something and I need to point out it does
work on species that have increased burnmods. Plasmamen self-cremating
themselves is actually a thing that happens in-game

* Removes fire cremation

* wew

---------

Co-authored-by: Time-Green <timkoster1@hotmail.com>
Co-authored-by: Gandalf <9026500+Gandalf2k15@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>