Fixes research.dmm APC Placement [MDB IGNORE] (#20194)
* Fixes research.dmm APC Placement (#74304)

## About The Pull Request

This PR was initially made to resolve the following issue:

```txt
[2023-03-27 14:53:51.874] Duplicate APC created at Research Genetics Research (123,126,12) /area/awaymission/research/interior/genetics. Original at Research Genetics Research (123,142,12) /area/awaymission/research/interior/genetics.
```

However, the APC placements in this entire away mission just sucked with
horrible area placement to accomodate for it. None of the areas here
have sprites, so it was pretty hard to piece out. Should be much better
now.
## Why It's Good For The Game
duplicate apc bad! better apc placement overall, it's more intuitive

## Changelog
it doesn't really matter as far as players are concerned

* Fixes research.dmm APC Placement

---------

Co-authored-by: san7890 <the@san7890.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>