Telescreen Updating Fix [MDB IGNORE] (#20263)
* Telescreen Updating Fix (#74424)

## About The Pull Request
Fixes https://github.com/tgstation/tgstation/issues/58843

Fixes the issue of only static being shown when closing the telescreen
while looking at a camera, then reopening the telescreen and selecting
the camera. Removes the static properly when reopening the telescreen,
instead of skipping it.
## Why It's Good For The Game

For most telescreens it wasn't too big of an issue, since it would just
mean reloading it by switching to a working camera, then back to the
desired camera. However, for telescreens connected to networks with only
one camera this isn't possible, meaning the telescreens become unusable
after closing it.
## Changelog
:cl:
fix: fixes telescreens showing static when reopening a camera
/:cl:

* Telescreen Updating Fix

---------

Co-authored-by: Ical <86125936+Ical92@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>