Adds cameras to emergency shuttles [MDB IGNORE] (#20268)
* Adds cameras to emergency shuttles (#74381)

## About The Pull Request

Adds cameras to each of the five emergency shuttles, which are linked to
the consoles in the bridge of each, as well as adding a telescreen to
the KiloStation shuttle to view the new cameras.
### Mapping March
Ckey to receive rewards: N/A

## Why It's Good For The Game

The emergency shuttles for the five stations all had consoles to view
cameras (other than KiloStation), but no cameras of their own, and were
not properly linked to any other camera network. Now the shuttles have
their own camera network that can be accessed properly from the shuttle
console.
## Changelog
:cl:
add: Added security cameras to the five "default" shuttles
add: Added camera telescreen to KiloStation Emergency Shuttle
/:cl:

* Adds cameras to emergency shuttles

---------

Co-authored-by: Ical <86125936+Ical92@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>