Increase first connect responsiveness by delaying vox sound preload [MDB IGNORE] (#20396)
* Increase first connect responsiveness by delaying vox sound preload (#74559)

These were getting sent too early and clogged up the queue of resources
to send the client, delaying chat load if the client had no byond rsc
cache and the cdn was not enabled.

Also this way if the cdn rsc download happens quick enough than these
calls can become noops. (if byond isn't smart enough to do that now, it
will be.)

* Increase first connect responsiveness by delaying vox sound preload

---------

Co-authored-by: Kyle Spier-Swenson <kyleshome@gmail.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>