You no longer drop the items in your pockets when your body is randomized [MDB IGNORE] (#20451)
* You no longer drop the items in your pockets when your body is randomized (#74581)

## About The Pull Request

When you are randomized by things like Mulligans or, in the case that
allowed me to find this bug, abductor transformation glands, the
contents on your pockets drop to the floor. That's because the proc that
checks if the items on your pockets fit post-transformation fails if
there is anything in your pocket... including the original item. It now
accounts for that and keeps your belongings safely inside your pockets
## Why It's Good For The Game

Fixes a bug. Small, but an annoying one.
## Changelog
:cl:
fix: the contents of your pockets no longer drop onto the floor when
your body is randomized
/:cl:

* You no longer drop the items in your pockets when your body is randomized

---------

Co-authored-by: A miscellaneous Fern <80640114+FernandoJ8@users.noreply.github.com>
Signed-off-by: Vladimir Veisman <v.veisman@flashie.me>